+++ src/csvcubed/inspect/inspectdatasetmanager.py
@@ -16,7 +16,7 @@ from csvcubed.cli.error_mapping import friendly_error_mapping
 from csvcubed.models.csvcubedexception import CsvToDataFrameLoadFailedException
 from csvcubed.models.csvwtype import CSVWType
 from csvcubed.models.cube.cube_shape import CubeShape
-from csvcubed.models.inspectdataframeresults import (
+from csvcubed.models.inspect.inspectdataframeresults import (
     CodelistHierarchyInfoResult,
     DatasetObservationsByMeasureUnitInfoResult,
     DatasetObservationsInfoResult,
