+++ src/csvcubed/inspect/sparql_handler/sparql_queries/ask_is_codelist.sparql
@@ -1,4 +1,4 @@
-ASK 
+ASK
 WHERE {
     ?conceptScheme a <http://www.w3.org/2004/02/skos/core#ConceptScheme>.
 }
