+++ src/csvcubed/inspect/sparql_handler/sparql_queries/ask_is_qb_dataset.sparql
@@ -1,4 +1,4 @@
-ASK 
+ASK
 WHERE {
     ?qbDataSet a <http://purl.org/linked-data/cube#DataSet>.
 }
