+++ src/csvcubed/inspect/sparql_handler/sparql_queries/select_catalog_metadata.sparql
