+++ src/csvcubed/inspect/sparql_handler/sparql_queries/select_column_definitions.sparql
