+++ src/csvcubed/inspect/sparql_handler/sparql_queries/select_is_pivoted_shape_data_set.sparql
