+++ src/csvcubed/inspect/sparql_handler/sparql_queries/select_labels_for_resource_uris.sparql
