+++ src/csvcubed/inspect/sparql_handler/sparql_queries/select_table_schema_properties.sparql
