+++ src/csvcubed/inspect/tableschema.py
@@ -4,6 +4,7 @@ Table Schema
 
 Provides functionality for handling table schema related features.
 """
+import json
 import logging
 from dataclasses import dataclass, field
 from functools import cached_property
@@ -12,23 +13,25 @@ from typing import Union
 from urllib.parse import urljoin
 
 import rdflib
+from rdflib import Graph
 from rdflib.util import guess_format
 
+from csvcubed.inspect.sparql_handler.csvw_repository import CsvWRepository
+from csvcubed.inspect.sparql_handler.sparql import path_to_file_uri_for_rdflib
+from csvcubed.inspect.sparql_handler.sparqlquerymanager import (
+    select_csvw_table_schema_file_dependencies,
+    select_metadata_dependencies,
+)
 from csvcubed.models.csvcubedexception import (
     FailedToLoadRDFGraphException,
     FailedToLoadTableSchemaIntoRdfGraphException,
     FailedToReadCsvwFileContentException,
     InvalidCsvwFileContentException,
 )
-from csvcubed.utils.csvw import load_table_schema_file_to_graph
 from csvcubed.utils.rdf import parse_graph_retain_relative
-from csvcubed.utils.sparql_handler.csvw_repository import CsvWRepository
-from csvcubed.utils.sparql_handler.sparql import path_to_file_uri_for_rdflib
-from csvcubed.utils.sparql_handler.sparqlquerymanager import (
-    select_csvw_table_schema_file_dependencies,
-    select_metadata_dependencies,
-)
 from csvcubed.utils.uri import looks_like_uri
+
+from ..utils.json import load_json_document
 
 _logger = logging.getLogger(__name__)
 
@@ -248,3 +251,30 @@ def add_triples_for_file_dependencies(
                 d.data_dump = absolute_url
 
         dependencies_to_load += new_dependencies
+
+
+def load_table_schema_file_to_graph(
+    table_schema_file_path: Union[str, Path],
+    table_schema_file_identifier: str,
+    graph: Graph,
+) -> None:
+    """
+    Given a tableSchema file definition at :obj:`table_schema_file_path`,
+     load the metadata as CSV-W flavoured RDF into the graph :obj:`graph`.
+    """
+    table_schema_document = load_json_document(table_schema_file_path)
+
+    # > When a schema is referenced by URL, this URL becomes the value of the @id property in the
+    # > normalized schema description, and thus the value of the schema annotation on the table.
+    #
+    # https://www.w3.org/TR/2015/REC-tabular-metadata-20151217/#table-schema
+    table_schema_document["@id"] = table_schema_file_identifier
+
+    # Provide the context to help generate all of the necessary RDF.
+    table_schema_document["@context"] = "http://www.w3.org/ns/csvw"
+
+    table_schema_document_json = json.dumps(table_schema_document)
+
+    parse_graph_retain_relative(
+        data=table_schema_document_json, format="json-ld", graph=graph
+    )
