+++ src/csvcubed/models/inspect/inspectdataframeresults.py
