+++ tests/unit/inspect/sparqlhandler/test_sparqlquerymanager.py
@@ -3,13 +3,13 @@ from pathlib import Path
 import pytest
 from rdflib import DCAT, RDF, RDFS, ConjunctiveGraph, Graph, Literal, URIRef
 
-from csvcubed.models.sparqlresults import MetadataDependenciesResult
-from csvcubed.utils.rdf import parse_graph_retain_relative
-from csvcubed.utils.sparql_handler.sparqlquerymanager import (
+from csvcubed.inspect.sparql_handler.sparqlquerymanager import (
     select_csvw_table_schema_file_dependencies,
     select_metadata_dependencies,
 )
-from csvcubed.utils.tableschema import add_triples_for_file_dependencies
+from csvcubed.inspect.tableschema import add_triples_for_file_dependencies
+from csvcubed.models.inspect.sparqlresults import MetadataDependenciesResult
+from csvcubed.utils.rdf import parse_graph_retain_relative
 from tests.helpers.repository_cache import (
     get_code_list_repository,
     get_csvw_rdf_manager,
