+++ tests/unit/utils/qb/test_components.py
@@ -2,12 +2,12 @@ from pathlib import Path
 
 import pytest
 
+from csvcubed.inspect.sparql_handler.sparql import path_to_file_uri_for_rdflib
 from csvcubed.utils.qb.components import (
     ComponentPropertyType,
     get_component_property_as_relative_path,
     get_component_property_type,
 )
-from csvcubed.utils.sparql_handler.sparql import path_to_file_uri_for_rdflib
 
 
 def test_printable_component_property_url():
